+++ Source/Plugins/VulkanRHI/Private/Plugins/VulkanRHI/PipelineState/VulkanPipelineState.cpp
@@ -101,9 +101,9 @@ namespace ob::rhi {
 
 
 		vk::PipelineVertexInputStateCreateInfo vertexInputInfo;
-        vertexInputInfo.vertexBindingDescriptionCount = bindingDescriptions.size();
+        vertexInputInfo.vertexBindingDescriptionCount = static_cast<uint32_t>(bindingDescriptions.size());
         vertexInputInfo.pVertexBindingDescriptions = bindingDescriptions.data();
-        vertexInputInfo.vertexAttributeDescriptionCount = attributeDescriptions.size();
+        vertexInputInfo.vertexAttributeDescriptionCount = static_cast<uint32_t>(attributeDescriptions.size());
         vertexInputInfo.pVertexAttributeDescriptions = attributeDescriptions.data();
 
 		vk::PipelineInputAssemblyStateCreateInfo inputAssemblyInfo;
