+++ Source/Test/Framework/Core/Source/Geometory/Size_Test.cpp
@@ -14,8 +14,8 @@ TEST(Size, Constructor) {
 	}
 	// 要素指定
 	{
-		f32 width = 1;
-		f32 height = 2;
+		s32 width = 1;
+		s32 height = 2;
 		Size size(width, height);
 		EXPECT_EQ(size.width, width);
 		EXPECT_EQ(size.height, height);
@@ -23,9 +23,9 @@ TEST(Size, Constructor) {
 	}
 	// 要素指定(depth)
 	{
-		f32 width = 1.0f;
-		f32 height = 2.0f;
-		f32 depth = 3.0f;
+		s32 width = 1;
+		s32 height = 2;
+		s32 depth = 3;
 		Size size(width, height, depth);
 		EXPECT_EQ(size.width, width);
 		EXPECT_EQ(size.height, height);
@@ -36,34 +36,34 @@ TEST(Size, Constructor) {
 TEST(Size,Operator) {
 	// ==
 	{
-		Size size1(1.0f, 2.0f);
-		Size size2(1.0f, 2.0f);
+		Size size1(1, 2);
+		Size size2(1, 2);
 		EXPECT_TRUE(size1 == size2);
 	}
 	{
-		Size size1(1.0f, 2.0f);
-		Size size2(2.0f, 2.0f);
+		Size size1(1, 2);
+		Size size2(2, 2);
 		EXPECT_FALSE(size1 == size2);
 	}
 	{
-		Size size1(1.0f, 2.0f);
-		Size size2(1.0f, 3.0f);
+		Size size1(1, 2);
+		Size size2(1, 3);
 		EXPECT_FALSE(size1 == size2);
 	}
 	// !=
 	{
-		Size size1(1.0f, 2.0f);
-		Size size2(1.0f, 2.0f);
+		Size size1(1, 2);
+		Size size2(1, 2);
 		EXPECT_FALSE(size1 != size2);
 	}
 	{
-		Size size1(1.0f, 2.0f);
-		Size size2(2.0f, 2.0f);
+		Size size1(1, 2);
+		Size size2(2, 2);
 		EXPECT_TRUE(size1 != size2);
 	}
 	{
-		Size size1(1.0f, 2.0f);
-		Size size2(1.0f, 3.0f);
+		Size size1(1, 2);
+		Size size2(1, 3);
 		EXPECT_TRUE(size1 != size2);
 	}
 }
@@ -71,30 +71,30 @@ TEST(Size,Operator) {
 TEST(Size, Getter) {
 	// area
 	{
-		Size size(1.0f, 2.0f);
-		EXPECT_EQ(size.area(), 2.0f);
+		Size size(1, 2);
+		EXPECT_EQ(size.area(), 2);
 	}
 	// volume
 	{
-		Size size(1.0f, 2.0f, 3.0f);
-		EXPECT_EQ(size.volume(), 6.0f);
+		Size size(1, 2, 3);
+		EXPECT_EQ(size.volume(), 6);
 	}
 }
 
 TEST(Size, Normalize) {
 	// normalize
 	{
-		Size size(1.0f, 2.0f, 3.0f);
+		Size size(1, 2, 3);
 		size.normalize();
-		EXPECT_EQ(size.width, 1.0f);
-		EXPECT_EQ(size.height, 2.0f);
-		EXPECT_EQ(size.depth, 3.0f);
+		EXPECT_EQ(size.width, 1);
+		EXPECT_EQ(size.height, 2);
+		EXPECT_EQ(size.depth, 3);
 	}
 }
 
 TEST(Size, Member) {
-	Size size(1.0f, 2.0f, 3.0f);
-	EXPECT_EQ(size.width, 1.0f);
-	EXPECT_EQ(size.height, 2.0f);
-	EXPECT_EQ(size.depth, 3.0f);
+	Size size(1, 2, 3);
+	EXPECT_EQ(size.width, 1);
+	EXPECT_EQ(size.height, 2);
+	EXPECT_EQ(size.depth, 3);
 }
